requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
streamlit
openai==0.27.4 # 1.30.1
python-dotenv
pdf2image
pillow
text
python-docx # for docs